+++ frontend/web/src/components/organisms/AllRooms.tsx
@@ -14,7 +14,7 @@ export const AllRooms = () => {
         <Stack divider={<StackDivider />} spacing="4">
           {rooms.map((room) => (
             <RoomCard
-              key={room.roomName}
+              key={room.roomId}
               roomName={room.roomName}
               lastUpdate={room.lastUpdate}
               memberAmount={room.memberAmount}
